Refactor Grafos
