+++ Main.py
@@ -1,43 +1,43 @@
-import Grafos
+import pgrafos
 
-grafo = Grafos.Grafo.generar_DorogovtsevMendes(30)
+grafo = pgrafos.Grafo.generar_DorogovtsevMendes(30)
 grafo.guardar("DorogovtsevMendes_30")
-grafo = Grafos.Grafo.generar_DorogovtsevMendes(100)
+grafo = pgrafos.Grafo.generar_DorogovtsevMendes(100)
 grafo.guardar("DorogovtsevMendes_100")
-grafo = Grafos.Grafo.generar_DorogovtsevMendes(500)
+grafo = pgrafos.Grafo.generar_DorogovtsevMendes(500)
 grafo.guardar("DorogovtsevMendes_500", "grafoPrueba")
 
-grafo = Grafos.Grafo.generar_BarbasiAlbert_variante(30, 4)
+grafo = pgrafos.Grafo.generar_BarbasiAlbert_variante(30, 4)
 grafo.guardar("BarbasiAlbert_variante_30")
-grafo = Grafos.Grafo.generar_BarbasiAlbert_variante(100, 4)
+grafo = pgrafos.Grafo.generar_BarbasiAlbert_variante(100, 4)
 grafo.guardar("BarbasiAlbert_variante_100")
-grafo = Grafos.Grafo.generar_BarbasiAlbert_variante(500, 4)
+grafo = pgrafos.Grafo.generar_BarbasiAlbert_variante(500, 4)
 grafo.guardar("BarbasiAlbert_variante_500")
 
-grafo = Grafos.Grafo.generar_ErdosRenyi(30, 40)
+grafo = pgrafos.Grafo.generar_ErdosRenyi(30, 40)
 grafo.guardar("ErdosRenyi_30")
-grafo = Grafos.Grafo.generar_ErdosRenyi(100, 40)
+grafo = pgrafos.Grafo.generar_ErdosRenyi(100, 40)
 grafo.guardar("ErdosRenyi_100")
-grafo = Grafos.Grafo.generar_ErdosRenyi(500, 40)
+grafo = pgrafos.Grafo.generar_ErdosRenyi(500, 40)
 grafo.guardar("ErdosRenyi_500")
 
-grafo = Grafos.Grafo.generar_geo_simple(30, 0.3)
+grafo = pgrafos.Grafo.generar_geo_simple(30, 0.3)
 grafo.guardar("geografico_30")
-grafo = Grafos.Grafo.generar_geo_simple(100, 0.3)
+grafo = pgrafos.Grafo.generar_geo_simple(100, 0.3)
 grafo.guardar("geografico_100")
-grafo = Grafos.Grafo.generar_geo_simple(500, 0.3)
+grafo = pgrafos.Grafo.generar_geo_simple(500, 0.3)
 grafo.guardar("geografico_500")
 
-grafo = Grafos.Grafo.generar_Gilbert(30, 0.3)
+grafo = pgrafos.Grafo.generar_Gilbert(30, 0.3)
 grafo.guardar("gilbert_30")
-grafo = Grafos.Grafo.generar_Gilbert(100, 0.3)
+grafo = pgrafos.Grafo.generar_Gilbert(100, 0.3)
 grafo.guardar("gilbert_100")
-grafo = Grafos.Grafo.generar_Gilbert(500, 0.3)
+grafo = pgrafos.Grafo.generar_Gilbert(500, 0.3)
 grafo.guardar("gilbert_500")
 
-grafo = Grafos.Grafo.generar_malla(10, 3)
+grafo = pgrafos.Grafo.generar_malla(10, 3)
 grafo.guardar("malla_30")
-grafo = Grafos.Grafo.generar_malla(10, 10)
+grafo = pgrafos.Grafo.generar_malla(10, 10)
 grafo.guardar("malla_100")
-grafo = Grafos.Grafo.generar_malla(50, 10)
+grafo = pgrafos.Grafo.generar_malla(50, 10)
 grafo.guardar("malla_500")
